Start implementing Quiz generation using TemplateParser
As noted in T152293 TemplateParser should be used instead of raw HTML.
This first commit will start using TemplateParser for a part of the Quiz
element. It is still needed to transform into templates the Questions and
the Setting element. As a result of this initial conversion I was able to drop
the use of an addittional CSS class (hideCorrection).

Bug: T152293
Change-Id: I7d1ae881d6d2c5b583509d5fa2e7f758715b13d5