Generate questions HTML via TemplateParser
This commit will move the generation of the Quiz's questions HTML to TemplateParser. Some work is still needed
here, especially it would be good to have CSS classes for the table style instead of inline CSS. I also added
some comments here and there especially for an unknown and undocumented feature.

Still to port to TemplateParser: Settings table, answers.
Bug: T152293

Change-Id: I557803731645f925db43166ae0d90eb1db8bdea8