Remove BadMethodCallException when creating quiz with no answer and zero
The quiz throws BadMethodCallException when no answer is provided or zero
is provided as answer, earlier it just showed syntax error.
This change creates templateParser before the if/else block and removes legacy
method empty.Added two tests one for empty value for question inside quiz and other
when 0 is answer.

Bug: T170733
Change-Id: I6a1fcf8a3707305dd8d41d93c4597950ba537fb4