Don't crash if quiz attempts to include a bad title
This fixes a regression introduced in mediawiki core by
99104944e884754fd39a3aabc55f721dbe322d54, which added stricter type
checks to Parser::fetchTemplate() and Parser::fetchTemplateAndTitle().
Quiz had been passing null here, which would cause `fetchTemplateAndTitle`
to return `[false, null]` which would then try to do a regexp match on
boolean `false` which would silently convert it to the empty string
and so the match would fail and you'd return the empty string.

In any case, now we'll explicitly check for a bad title and return an
empty string directly.  It might perhaps be better to emit an error
message of some sort, but in general there's not a precedent for that
in the Quiz extension -- removing bad input from the output seems to
be its modus operandi -- and beside it would be better to separate out
this regression fix from any attempt to reform the behavior of Quiz.

Bug: T251409
Change-Id: Ib82b5cedd711cc0e284412251784d02ca9938696
(cherry picked from commit 1513c7082d861b612d48ee793542643e3516f8a3)