Remove test that does not test anything
Ensuring that a certain class has a static property with a certain name
is only verifying an implementation detail of that class, which the test
should not do. The test itself does not provide any value.

This is needed for PHPUnit 10 compatibility, where the
assertClassHasStaticAttribute method has been removed with no
replacement.

Bug: T342110
Change-Id: I23a9a144968bf77b9800649dcc391e516c6a7fae