refactor: Replace dynamic method call with static one
The dynamic method call used only information from the object the
methods were called on, so why not process the information in there?
Especially as the information is also accessed directly in
Question::basicTypeParseObject(), which could cause discrepancy
(what if someone called Question::multipleChoiceParseObject() while
Question::mType was set to `singleChoice`?). And two of the three
dynamically called methods were one-liner wrappers, so they could just
be inlined.

Removing public methods from an object passed to a hook is technically
a breaking change, but I don’t think anything actually uses that hook
since QuizTabulate was archived in T274045, so I hope this is fine.

Change-Id: I60fbe2047fa8ae63ed72464155eca403229489db