Try to help Parsoid avoid emitting duplicate ids
It was reported that the Quiz extension was producing duplicate ids,
which is discouraged, on pages with multiple quizzes,
https://www.mediawiki.org/wiki/Special:Diff/7798148
https://www.mediawiki.org/wiki/Parsoid/Parser_Unification/Instructions_for_editors#Do_not_use_duplicate_ids

Some of the ids seem to exist on input elements to match up with labels.
Those are appended with the quiz id to make them unique per page.

However, the quiz id itself is generated from a counter that was being
reset after each parse.  For extensions where Parsoid doesn't have a
native implementation, it calls out to the legacy parser to parse each
instance of the tag on the page separately.  The counter would reset
each time so all instances had the same quiz id.  That issue isn't
present with the legacy parser.

The fix here is to not reset the counter.  It's a bit unclear why the
counter was being reset to begin with.  In the initial implementation in
4c00e1d5 the styles were inlined and only inserted for quiz id 0.  So
maybe the thought was that each invocation of the parser was going to
need the counter reset to include the styles.  Resetting the counter was
not without issue since it risks having multiple quizzes on the page
with the same id, as in T152398.

In any case, these days, addModule(s|Styles) deduplicate modules so we
can call that at each invocation.  And the counter will reset when the
request terminates.

Also, getQuizId() was removed as a fishy instance method that was
returning static state that changes when quizzes are constructed.

Bug: T272944
Change-Id: I9df44b64e6796260f88b74f7cca868505fa00b08